admin/requirements.txt
http.client
json
tkinter
pillow
requests
matplotlib
